+++ packages/kit/src/runtime/client/client.js
@@ -221,6 +221,9 @@ export function create_client(app, target) {
 			if (navigation_result.type === 'redirect') {
 				return goto(new URL(navigation_result.location, url).href, {}, [url.pathname], nav_token);
 			} else {
+				if (navigation_result.props.page !== undefined) {
+					page = navigation_result.props.page;
+				}
 				root.$set(navigation_result.props);
 			}
 		}
@@ -1175,7 +1178,7 @@ export function create_client(app, target) {
 			document.activeElement !== document.body;
 
 		if (!keepfocus && !changed_focus) {
-			await reset_focus();
+			reset_focus();
 		}
 
 		autoscroll = true;
@@ -1185,6 +1188,11 @@ export function create_client(app, target) {
 		}
 
 		navigating = false;
+
+		if (type === 'popstate') {
+			restore_snapshot(current_history_index);
+		}
+
 		callbacks.after_navigate.forEach((fn) =>
 			fn(/** @type {import('types').AfterNavigate} */ (navigation))
 		);
@@ -1770,7 +1778,6 @@ export function create_client(app, target) {
 					}
 
 					const delta = history_index - current_history_index;
-					let blocked = false;
 
 					await navigate({
 						url: new URL(event.state[PAGE_URL_KEY] ?? location.href),
@@ -1785,17 +1792,10 @@ export function create_client(app, target) {
 						},
 						blocked: () => {
 							history.go(-delta);
-							blocked = true;
 						},
 						type: 'popstate',
 						delta
 					});
-
-					if (!blocked) {
-						restore_snapshot(current_navigation_index);
-					}
-				} else {
-					console.log('popstate to a non-SvelteKit index');
 				}
 			});
 
@@ -1880,14 +1880,30 @@ export function create_client(app, target) {
 					});
 				});
 
+				/** @type {Array<import('./types').BranchNode | undefined>} */
+				const branch = await Promise.all(branch_promises);
+
+				const parsed_route = routes.find(({ id }) => id === route.id);
+
+				// server-side will have compacted the branch, reinstate empty slots
+				// so that error boundaries can be lined up correctly
+				if (parsed_route) {
+					const layouts = parsed_route.layouts;
+					for (let i = 0; i < layouts.length; i++) {
+						if (!layouts[i]) {
+							branch.splice(i, 0, undefined);
+						}
+					}
+				}
+
 				result = await get_navigation_result_from_branch({
 					url,
 					params,
-					branch: await Promise.all(branch_promises),
+					branch,
 					status,
 					error,
 					form,
-					route: routes.find(({ id }) => id === route.id) ?? null
+					route: parsed_route ?? null
 				});
 			} catch (error) {
 				if (error instanceof Redirect) {
@@ -2046,12 +2062,44 @@ function reset_focus() {
 			root.removeAttribute('tabindex');
 		}
 
-		return new Promise((resolve) => {
+		// capture current selection, so we can compare the state after
+		// snapshot restoration and afterNavigate callbacks have run
+		const selection = getSelection();
+
+		if (selection && selection.type !== 'None') {
+			/** @type {Range[]} */
+			const ranges = [];
+
+			for (let i = 0; i < selection.rangeCount; i += 1) {
+				ranges.push(selection.getRangeAt(i));
+			}
+
 			setTimeout(() => {
+				if (selection.rangeCount !== ranges.length) return;
+
+				for (let i = 0; i < selection.rangeCount; i += 1) {
+					const a = ranges[i];
+					const b = selection.getRangeAt(i);
+
+					// we need to do a deep comparison rather than just `a !== b` because
+					// Safari behaves differently to other browsers
+					if (
+						a.commonAncestorContainer !== b.commonAncestorContainer ||
+						a.startContainer !== b.startContainer ||
+						a.endContainer !== b.endContainer ||
+						a.startOffset !== b.startOffset ||
+						a.endOffset !== b.endOffset
+					) {
+						return;
+					}
+				}
+
+				// if the selection hasn't changed (as a result of an element being (auto)focused,
+				// or a programmatic selection, we reset everything as part of the navigation)
 				// fixes https://github.com/sveltejs/kit/issues/8439
-				resolve(getSelection()?.removeAllRanges());
+				selection.removeAllRanges();
 			});
-		});
+		}
 	}
 }